APP CHAT GPT/requirements.txt

deepface
tensorflow==2.18.0
tf-keras
pillow
opencv-python
